resolved PR comments

+++ src/gps_module/gps_module.h
@@ -8,6 +8,8 @@
 #ifndef GPS_MODULE_H_
 #define GPS_MODULE_H_
 
+#include <stdint.h>
+
 #define DEBUG
 
 /* gps_data_struct struct to holds GPS data*/
@@ -29,6 +31,6 @@ void *read_from_gps(void *); /* pthread to handle gps read */
 void get_lat_log(double *);
 void get_dops(char **, char *);
 void get_gps_data(char *, struct gps_data_struct *);
-void get_gps_param_by_position(char **,char*,int);
+void get_gps_param_by_position(char **,char*,uint8_t);
 
 #endif
